rename plots to dataset_plots
